requirements.txt
# Core ML / DL
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.16
numpy>=1.26.0
scipy>=1.13.0

# CT / Medical imaging
pydicom>=2.4.4
nibabel>=5.2.1
Pillow>=10.3.0

# Image augmentation
albumentations>=1.4.3

# Data & features
pandas>=2.2.2
scikit-learn>=1.4.2
PyYAML>=6.0.1

# Experiment tracking
mlflow>=2.11.3
dvc>=3.50.1

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# Visualisation
matplotlib>=3.8.4
seaborn>=0.13.2
